preparation for chart

+++ app/src/main/java/com/evgenii/coolgraph/domain/PointResponse.kt
@@ -3,4 +3,4 @@ package com.evgenii.coolgraph.domain
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PointResponse(val x: Double, val y: Double)
+data class PointResponse(val x: Float, val y: Float)
